Fix endpoint
