+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/routing/api/ReportRouteConfig.kt
@@ -40,11 +40,6 @@ object ReportRouteConfig : RouteConfig
             APIEndpoint("/reports/:name/latest/changelog/", ReportController::class, "getLatestChangelogByName")
                     .json()
                     .transform()
-                    .secure(readReports),
-            APIEndpoint("/report/:name/dependencies/",
-                   ReportController::class, "getDependencies")
-                    .json()
-                    .transform()
                     .secure(readReports)
     )
 }
