+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/controllers/web/ReportController.kt
@@ -137,4 +137,11 @@ class ReportController(
         reportRepository.publish(ids)
         return okayResponse()
     }
+
+    fun getDependencies(): String
+    {
+        val name = context.params(":name")
+        val response = orderlyServerAPI.get("/v1/reports/$name/dependencies/", context)
+        return passThroughResponse(response)
+    }
 }
